Add first implementation
Note that storage and context are not implemented. This is left for the
application side.

A standard implementation for the context would be easy to add here,
ensuring it can be easily extended or replaced.

The storage is too dependent on the application and should therefore remain
abstract in this library.